requirements.txt
flask
flask-sqlalchemy
flask-restful
flask-marshmallow
marshmallow-sqlalchemy
sqlalchemy-utils
passlib
